Start Wayland server, add seat and dmabuf handler

+++ src/flutter_engine/wayland_messages.rs
@@ -6,7 +6,7 @@ use smithay::wayland::shell::xdg;
 use crate::flutter_engine::platform_channels::encodable_value::EncodableValue;
 
 pub struct SurfaceCommitMessage {
-    pub view_id: usize,
+    pub view_id: u64,
     pub role: Option<&'static str>,
     pub texture_id: i32,
     pub buffer_delta: Option<Point<i32, Logical>>,
